xhtml build is now working on View/Compiler API instead of custom implementation in builder https://online.sbis.ru/opendoc.html?guid=7815e62f-bdde-437c-81e6-e9dc81f4ca13
