+++ test/build-template.test.js
@@ -37,4 +37,36 @@ describe('build template', () => {
       );
       testResults(result);
    });
+   it('basic tmpl', async() => {
+      let localization = true;
+      const testResults = (result) => {
+         result.text.startsWith('define(\'wml!TestModule/TestWml\'').should.equal(true);
+         result.nodeName.should.equal('wml!TestModule/TestWml');
+         result.config.should.deep.equal({
+            fileName: 'TestModule/TestWml.wml',
+            fromBuilderTmpl: true,
+            createResultDictionary: true,
+            componentsProperties: 'path/to/components-properties.json',
+            generateCodeForTranslations: localization
+         });
+      };
+
+      let result = await processingTmpl.buildTmpl(
+         '<div>{{1+1}}</div>',
+         path.normalize('TestModule/TestWml.wml'),
+         'path/to/components-properties.json',
+         localization
+      );
+      testResults(result);
+
+      // disable localization, after new build localization should be disabled in result
+      localization = false;
+      result = await processingTmpl.buildTmpl(
+         '<div>{{1+1}}</div>',
+         path.normalize('TestModule/TestWml.wml'),
+         'path/to/components-properties.json',
+         localization
+      );
+      testResults(result);
+   });
 });
